Reorder (and restructure) the sections in the README
Previously, the sections were
* Supported syntax
* Module Aliases
* New functions, macros, and methods
* Renaming
* New macros
* Other changes
* New types
* Developer tips

This changes the order to
* Supported syntax
* New functions and methods
* New macros
* New types
* Renaming
* Other changes
* Developer tips

getting rid of the ambiguity where to put new macros and removing the
"Module Aliases" section (if needed, these could probably go into
"Renaming").

Also, the entry from #668 is moved from "Other changes" to "New
functions and methods" as it is a new method.

Finally, presently empty sections get a "None." placeholder entry.